+++ eventregistry/Analytics.py
@@ -98,24 +98,34 @@ class Analytics:
         return self._er.jsonRequestAnalytics("/api/v1/ner", {"text": text})
 
 
-    def trainTopicOnTweets(self, twitterQuery, useTweetText = True, maxConcepts = 20, maxCategories = 10, maxTweets = 2000, notifyEmailAddress = None):
+    def trainTopicOnTweets(self, twitterQuery, useTweetText=True, useIdfNormalization=True,
+            normalization="linear", maxTweets=2000, maxUsedLinks=500, ignoreConceptTypes=[],
+            maxConcepts = 20, maxCategories = 10, notifyEmailAddress = None):
         """
         create a new topic and train it using the tweets that match the twitterQuery
         @param twitterQuery: string containing the content to search for. It can be a Twitter user account (using "@" prefix or user's Twitter url),
                 a hash tag (using "#" prefix) or a regular keyword.
         @param useTweetText: do you want to analyze the content of the tweets and extract the concepts mentioned in them? If False, only content shared
             in the articles in the user's tweets will be analyzed
+        @param useIdfNormalization: normalize identified concepts by their IDF in the news (punish very common concepts)
+        @param normalization: way to normalize the concept weights ("none", "linear")
+        @param maxTweets: maximum number of tweets to collect (default 2000, max 5000)
+        @param maxUsedLinks: maximum number of article links in the tweets to analyze (default 500, max 2000)
+        @param ignoreConceptTypes: what types of concepts you would like to ignore in the profile. options: person, org, loc, wiki or an array with those
         @param maxConcepts: the number of concepts to save in the final topic
         @param maxCategories: the number of categories to save in the final topic
         @param maxTweets: the maximum number of tweets to collect for the user to analyze
         @param notifyEmailAddress: when finished, should we send a notification email to this address?
         """
         assert maxTweets < 5000, "we can analyze at most 5000 tweets"
-        params = {"twitterQuery": twitterQuery,
-            "useTweetText": useTweetText, "maxConcepts": maxConcepts, "maxCategories": maxCategories,
-            "maxTweets": maxTweets}
+        params = {"twitterQuery": twitterQuery, "useTweetText": useTweetText,
+            "useIdfNormalization": useIdfNormalization, "normalization": normalization,
+            "maxTweets": maxTweets, "maxUsedLinks": maxUsedLinks,
+            "maxConcepts": maxConcepts, "maxCategories": maxCategories }
         if notifyEmailAddress:
             params["notifyEmailAddress"] = notifyEmailAddress
+        if len(ignoreConceptTypes) > 0:
+            params["ignoreConceptTypes"] = ignoreConceptTypes
         return self._er.jsonRequestAnalytics("/api/v1/trainTopicOnTwitter", params)
 
 
@@ -127,6 +137,14 @@ class Analytics:
         return self._er.jsonRequestAnalytics("/api/v1/trainTopic", { "action": "createTopic", "name": name})
 
 
+    def trainTopicClearTopic(self, uri):
+        """
+        if the topic is already existing, clear the definition of the topic. Use this if you want to retrain an existing topic
+        @param uri: uri of the topic (obtained by calling trainTopicCreateTopic method) to clear
+        """
+        return self._er.jsonRequestAnalytics("/api/v1/trainTopic", { "action": "clearTopic", "uri": uri })
+
+
     def trainTopicAddDocument(self, uri, text):
         """
         add the information extracted from the provided "text" to the topic with uri "uri"
@@ -136,22 +154,15 @@ class Analytics:
         return self._er.jsonRequestAnalytics("/api/v1/trainTopic", { "action": "addDocument", "uri": uri, "text": text})
 
 
-    def trainTopicFinishTraining(self, uri, maxConcepts = 20, maxCategories = 10, idfNormalization = True):
-        """
-        add the information extracted from the provided "text" to the topic with uri "uri"
-        @param uri: uri of the topic (obtained by calling trainTopicCreateTopic method)
-        @param maxConcepts: number of top concepts to save in the topic
-        @param maxCategories: number of top categories to save in the topic
-        @param idfNormalization: should the concepts be normalized by punishing the commonly mentioned concepts
-        @param returns: returns the trained topic: { concepts: [], categories: [] }
-        """
-        return self._er.jsonRequestAnalytics("/api/v1/trainTopic", {"action": "finishTraining", "uri": uri, "maxConcepts": maxConcepts, "maxCategories": maxCategories, "idfNormalization": idfNormalization})
-
-
-    def trainTopicGetTrainedTopic(self, uri):
+    def trainTopicGetTrainedTopic(self, uri, maxConcepts = 20, maxCategories = 10,
+            ignoreConceptTypes=[], idfNormalization = True):
         """
         retrieve topic for the topic for which you have already finished training
         @param uri: uri of the topic (obtained by calling trainTopicCreateTopic method)
+        @param maxConcepts: number of top concepts to retrieve in the topic
+        @param maxCategories: number of top categories to retrieve in the topic
+        @param ignoreConceptTypes: what types of concepts you would like to ignore in the profile. options: person, org, loc, wiki or an array with those
+        @param idfNormalization: should the concepts be normalized by punishing the commonly mentioned concepts
         @param returns: returns the trained topic: { concepts: [], categories: [] }
         """
-        return self._er.jsonRequestAnalytics("/api/v1/trainTopic", { "action": "getTrainedTopic", "uri": uri })
+        return self._er.jsonRequestAnalytics("/api/v1/trainTopic", { "action": "getTrainedTopic", "uri": uri, "maxConcepts": maxConcepts, "maxCategories": maxCategories, "idfNormalization": idfNormalization })